% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ECA11_ERAINT.Germany.JJA}
\alias{ECA11_ERAINT.Germany.JJA}
\title{Daily precipitation records used in this study range from 1979 to 2008 VALUE experimental
period, for a subset of 11 stations in southeast Germany extracted from 
the VALUE spatial validation experiment (A description of the experiment and
the data can be found in Widmann et al. (2019), \url{https://doi.org/10.1002/joc.6024}); and for 
6 grid points taken from the ERA-Interim dataset.
Precipitation has been discretized to 0/1 absence/presence of precipitation. Only June, 
July and August are included (JJA season).}
\format{An object of class \code{list} of length 2.}
\usage{
ECA11_ERAINT.Germany.JJA
}
\description{
Daily precipitation records used in this study range from 1979 to 2008 VALUE experimental
period, for a subset of 11 stations in southeast Germany extracted from 
the VALUE spatial validation experiment (A description of the experiment and
the data can be found in Widmann et al. (2019), \url{https://doi.org/10.1002/joc.6024}); and for 
6 grid points taken from the ERA-Interim dataset.
Precipitation has been discretized to 0/1 absence/presence of precipitation. Only June, 
July and August are included (JJA season).
}
\references{
The stations dataset is provided by the European Climate Assessment & Dataset
 project (ECA&D). For ERA-Interim, see \url{http://doi.org/10.1002/qj.828}
}
\author{
Mikel N. Legasa \email{mikel.legasa@unican.es}
}
\keyword{datasets}
