% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ECA53.Germany.DJF}
\alias{ECA53.Germany.DJF}
\title{Daily precipitation records used in this study range from 1979 to 2008 VALUE experimental
period, for the 53 stations in Germany extracted of the VALUE spatial validation experiment
 (A description of the experiment and the data can be found in Widmann et al. (2019),
 \url{https://doi.org/10.1002/joc.6024}). Precipitation has been discretized to 0/1 
 absence/presence of precipitation. Only December, January and February are included 
 (DJF season).}
\format{
An object of class \code{list} of length 5.
}
\usage{
data(ECA53.Germany.DJF)
}
\description{
Daily precipitation records used in this study range from 1979 to 2008 VALUE experimental
period, for the 53 stations in Germany extracted of the VALUE spatial validation experiment
 (A description of the experiment and the data can be found in Widmann et al. (2019),
 \url{https://doi.org/10.1002/joc.6024}). Precipitation has been discretized to 0/1 
 absence/presence of precipitation. Only December, January and February are included 
 (DJF season).
}
\references{
The dataset is provided by the European Climate Assessment & Dataset
 project (ECA&D).
}
\author{
Mikel N. Legasa \email{mikel.legasa@unican.es}
}
\keyword{datasets}
