% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateWeather.R
\name{generateWeather}
\alias{generateWeather}
\title{Generate multivariate weather series using a Bayesian Network Weather Generator.}
\usage{
generateWeather(
  wg,
  initial = NULL,
  n = 1,
  x = NULL,
  inference.type = NULL,
  initial.date = NULL,
  advance.type = "simulation",
  threshold.vector = 0.5,
  resample.size = 10000,
  event = "1"
)
}
\arguments{
\item{wg}{Bayesian network weather generator model, as output from \code{buildWeatherGenerator()}}

\item{initial}{An initial observation, as vector. If \code{NULL}, a random observation will be chosen.}

\item{n}{Number of observations to generate.}

\item{x}{A predictors object (grid)}

\item{inference.type}{Either \code{"approximate"} or \code{"exact"}. If junction tree is unavailable, exact inference is
deactivated}

\item{initial.date}{A date element. If provided, rownames will be labeled after this date.}

\item{advance.type}{\code{"simulation"} (default) is used to generate stochastic weather series. 
Other options are included for debugging purposes, which are \code{"exact"}, so that the series are generated
deterministically using most likely and \code{threshold.vector}.}

\item{threshold.vector}{Ignored if \code{advance.type = "simulation"}, used as the threshold for deciding
if probability value should be 1 or 0.}

\item{event}{Name of the positive event.}
}
\description{
Uses a Bayesian network model, output from \code{buildWeatherGenerator()}
 to simulate a multivariate series. If the model was learnt without predictors, x
 must be NULL. Otherwise x is required.
}
