% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelHandler.R
\name{parallelHandler}
\alias{parallelHandler}
\title{Parallelization Cluster Starter}
\usage{
parallelHandler(
  type,
  n.cores,
  PSOCK.funcExports.list = list(),
  PSOCK.varExports.list = list(),
  cl = NULL
)
}
\arguments{
\item{type}{Cluster type, either FORK or PSOCK.}
}
\description{
Parallelization Cluster Starter
}
\author{
M.N. Legasa
}
