% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCBN.R
\name{plotCBN}
\alias{plotCBN}
\title{Plot Bayesian Network Weather Model}
\usage{
plotCBN(
  CBN,
  title = NULL,
  dev.new = FALSE,
  no.labels = FALSE,
  vertex.label.dist = 0.8,
  no.colors = TRUE,
  node.size = 3,
  edge.width = 0.6,
  edge.arrow.size = 0.2,
  edges.color = c("azure3", "grey", "azure4", "black"),
  mark.edge.strength = FALSE,
  break.axis = 1,
  separation.ratio = 0.1,
  Nlabels = 4,
  return.plot = TRUE
)
}
\arguments{
\item{CBN}{Downscaling Bayesian Network, as returned 
by \code{buildWeatherGenerator()}, \code{buildPredictive()} or
\code{buildDescriptive()}. Uses \code{igraph} package.}

\item{title}{Plot title.}

\item{dev.new}{Plot in a new device.}

\item{no.labels}{Do not print labels}

\item{vertex.label.dist}{Distance from node to labels}

\item{no.colors}{Set to \code{TRUE} to remove all coloring.}

\item{return.plot}{Set to TRUE to return the plot as a variable.}

\item{...}{Refer to \code{igraph} package.}
}
\description{
Plot Bayesian Network Weather Model
}
\author{
Mikel N. Legasa
}
