% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareDataDynamicBN.R
\name{prepareDataDynamicBN}
\alias{prepareDataDynamicBN}
\title{Prepare Datasets for Dinamic Bayesian Network.}
\usage{
prepareDataDynamicBN(data_, epochs)
}
\arguments{
\item{data}{Expects output from \code{\link[downscaleR.BN]{prepare_predictors.forBN()}}.}
}
\description{
prepareDataDinamicBN()
}
\author{
MN Legasa
}
