% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBN.R
\name{queryBN}
\alias{queryBN}
\title{Query Bayesian Network Model.}
\usage{
queryBN(evidence, dbn, evidence.nodes, predictands, type = "exact",
  which_ = "marginal", resample.size = 10000, cl = NULL)
}
\arguments{
\item{evidence}{An evidence vector.}

\item{dbn}{A Bayesian network model, as output by \code{buildWeatherGenerator()},
\code{buildPredictive()} or \code{buildDescriptive()}}

\item{evidence.nodes}{Names of the nodes for which evidence will be provided. 
Must have the same length as \code{evidence}.}

\item{predictands}{Nodes queried.}

\item{type}{Inference type. Either \code{"simulation"}, \code{"approximate"}, \code{"exact"},
or \code{"random"}. Exact inference requires the compiled junction tree and \code{gRain} library. 
Approximate uses \code{cpdist()} function from \code{bnlearn}. Simulation is used for weather generation,
and generates stochastic predictions. \code{"random"} generates a non spatially consistent observation, 
based directly on the probability vector obtained using only evidence nodes.}

\item{which_}{Probabilities to return. Used only for exact inference.}

\item{resample.size}{Number of samples for the likelihood weighting approximate inference method.}

\item{cl}{A cluster object from \code{parallel} library, used only if \code{"approximate"} inference
is selected.}
}
\description{
Performs inference in a Bayesian network model, which can be output from 
\code{buildWeatherGenerator()}, \code{buildPredictive()} and \code{buildDescriptive()}.
}
