% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparePredictorsBN.R
\name{preparePredictorsBN}
\alias{preparePredictorsBN}
\title{Prepare datasets from \code{loadeR} for the BN.}
\usage{
preparePredictorsBN(grid, rm.na = TRUE, rm.na.mode = "observations")
}
\arguments{
\item{grid}{Expects grid from `loadeR`.}
}
\value{
A processed dataset to be passed to \code{buildPredictive()}.
}
\description{
preparePredictorsBN()
}
\author{
Mikel N. Legasa
}
